backend/requirements.txt
certifi==2024.7.4
Django==5.0.2
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
django-storages==1.14.3
gunicorn==21.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
supabase==2.4.3
urllib3==2.2.1
gevent==24.2.1
psycopg[c]==3.1.18
pillow==10.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0